pyproject.toml
[project]
name = "read-book-note"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]
